requirements.txt
streamlit>=1.32.0
pandas>=2.2.1
plotly>=5.19.0
fpdf2>=2.7.8
sqlite3-api
numpy>=1.26.4
